backend/requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0

# AI - Mistral
mistralai==1.5.0
httpx==0.25.2

# Vector database
chromadb==0.4.22
sentence-transformers==2.3.1

# NER / Anonymization
gliner==0.2.7

# Document processing
python-docx==1.1.0
PyMuPDF==1.24.0
openpyxl==3.1.2
Pillow==10.2.0

# Utils
python-dotenv==1.0.0
jinja2==3.1.3
aiofiles==23.2.1

# Rate limiting & security
slowapi==0.1.9
